requirements.txt

streamlit
transformers
pandas
torch
google-generativeai